refactor(agents-server-ui): comments visibility via live query, shared fork-from-here hook
GenericChatBody no longer filters comment rows in JS — useEntityTimeline
takes a comments flag and omits the comments extraSource when the view
hides them, so the query engine maintains the row set. The duplicated
fork-from-here anchor map moves to a shared useForkFromHere hook.
